refactor(st-09063): modularize multi-agent worker node
