+++ docs/st09063-multi-agent-worker-node-modularization.md
@@ -1,0 +1,57 @@
+# ST-09063: Multi-Agent Worker Node Modularization
+
+## Summary
+
+`packages/patterns/src/multi-agent/nodes/worker.ts` was reduced from `357` lines to a `140` line public worker implementation that stays below the `300` line planning cutoff while preserving the stable `createWorkerNode(...)` entrypoint behind `packages/patterns/src/multi-agent/nodes.ts`.
+
+The extracted runtime modules are each focused on one responsibility and also remain below the planning cutoff:
+
+- `packages/patterns/src/multi-agent/nodes/worker-model.ts` (`86` lines)
+- `packages/patterns/src/multi-agent/nodes/worker-workload.ts` (`134` lines)
+- `packages/patterns/src/multi-agent/nodes/worker-types.ts` (`10` lines)
+
+The public worker implementation still owns assignment lookup, execution-path selection, and the shared success/error orchestration, while model invocation/prompting and workload/error-result bookkeeping now live behind explicit internal seams.
+
+## Test Modularization
+
+The monolithic `packages/patterns/tests/multi-agent/nodes.test.ts` was replaced with a `6` line public entrypoint that imports focused suites:
+
+- `packages/patterns/tests/multi-agent/nodes/supervisor-routing.ts`
+- `packages/patterns/tests/multi-agent/nodes/supervisor-workload.ts`
+- `packages/patterns/tests/multi-agent/nodes/worker-core.ts`
+- `packages/patterns/tests/multi-agent/nodes/worker-workload.ts`
+- `packages/patterns/tests/multi-agent/nodes/worker-overrides.ts`
+- `packages/patterns/tests/multi-agent/nodes/aggregator-node.ts`
+- `packages/patterns/tests/multi-agent/nodes/shared.ts`
+
+This keeps the public test command stable while separating supervisor routing, supervisor workload, worker execution flow, worker workload ownership, worker override/merge behavior, and aggregator behavior into reviewable units.
+
+## Test Strategy And Compatibility
+
+This story is behavior-preserving modularization rather than a behavior change, so a literal failing test for "the worker file got smaller" would only assert repository structure. The practical test-first substitute was:
+
+1. extract the `createWorkerNode` assertions from `packages/patterns/tests/multi-agent/nodes.test.ts` into focused suites first
+2. run the unchanged public entrypoint
+3. modularize `packages/patterns/src/multi-agent/nodes/worker.ts` behind the same public facade
+4. re-run the same public entrypoint plus broader validation
+
+Compatibility notes:
+
+- `createWorkerNode` remains exported from `packages/patterns/src/multi-agent/nodes.ts`
+- existing worker-node behavior remains intact for custom execution, ReAct wrapping, model invocation, handoff propagation, workload decrement semantics, error-result handling, and public imports
+- no CI or workflow change was required because the story only reorganizes internal modules and test surfaces behind existing public entrypoints
+
+## Validation
+
+- `pnpm test --run packages/patterns/tests/multi-agent/nodes.test.ts`
+  - before production split: `1` file passed, `33` tests passed
+  - after production split: `1` file passed, `33` tests passed
+- `pnpm --filter @agentforge/patterns typecheck`
+- `pnpm lint:explicit-any:baseline`
+  - workspace: `84/289`
+  - patterns: `2/28`
+  - no baseline regression; follow-up baseline file update not included in this behavior-preserving refactor
+
+## Explicit-`any` Notes
+
+This story did not increase explicit-`any` usage. The touched multi-agent worker modules remain consistent with the existing baseline, and the baseline improvement notice from `pnpm lint:explicit-any:baseline` is deferred to a separate follow-up because the committed baseline file itself is outside this story's scope.
