+++ packages/patterns/src/multi-agent/nodes/worker-types.ts
@@ -1,0 +1,10 @@
+import type { MultiAgentStateType } from '../state.js';
+
+export interface InvokableWorkerModel {
+  invoke(messages: unknown): Promise<{ content: unknown }>;
+}
+
+export type TaskResultAndMessage = Pick<
+  MultiAgentStateType,
+  'completedTasks' | 'messages'
+>;
